Move Auth Code exchange logic into AuthenticationProvider
The exchange of the authorization code grant for an access token was originally
handled in AuthorizationCodeGrantHandler which was called by the AuthorizationCodeGrantProcessingFilter.
This logic really belongs in an AuthenticationProvider as we're exchanging
one credential  (auth code) for another (access token).

This commit removes AuthorizationCodeGrantHandler in favour of
AuthorizationGrantTokenExchanger which is ultimately called by the
AuthorizationCodeGrantAuthenticationProvider.

Fixes gh-18